refactor(context): revert to sync and async contexts
